+++ back-end/requirements.txt
@@ -1,7 +1,6 @@
 asgiref==3.5.2
 Django==4.1.1
 django-autoslug==1.9.8
-django-phonenumber-field==7.0.0
 django-rest-framework==0.1.0
 djangorestframework==3.14.0
 Pillow==9.2.0
